requirements.txt
openai-agents==0.2.8
python-dotenv==1.0.1
chromadb
openai==1.99.9
pandas
gspread
oauth2client
flask_cors
streamlit
openai-agents[litellm]